refactor(query): split the attach subsystem out of inspect.rs
I moved the multi-service log-attach subsystem — `AttachOutcome`,
`AttachOptions`, `AttachSummary`, `Engine::attach_logs`,
`Engine::attach_logs_with_options`, `Engine::attach_logs_with`, plus the
private `container_exit_code` helper, the `Phase` / `StreamEnd` enums,
`wait_for_phase` and the `attach_outcome_tests` module — out of
`internal/engine/query/inspect.rs` into a new
`internal/engine/query/attach.rs`. Pure relocation: no behaviour change,
no signature change, no test change.

| File | Before | After |
|---|---|---|
| `internal/engine/query/inspect.rs` | 517 | 225 |
| `internal/engine/query/attach.rs`  |  — | 299 |

Both files are now under the 500-line hard limit; `attach.rs` is also
under the 300-line soft limit.

Why a separate module: `top`, `port` and the single-service `attach`
command don't share any state with the abort path and don't need the
public types the abort path defines. Keeping them together forced
`inspect.rs` past the cap; splitting along the `attach_logs*` /
`attach`-only seam keeps each file under it.

The `attach_stream_tests` integration tests stay declared from
`query/mod.rs` (same place as before) and just import the types from
`super::attach` instead of `super::inspect`. `query/mod.rs` re-exports
the three types from `attach` instead of `inspect`, so the public
surface — `podup::AttachOptions`, `podup::AttachOutcome`,
`podup::AttachSummary` — is unchanged.

`cargo semver-checks check-release --baseline-version 4.0.0` still
reports 223 pass / 0 fail / 30 skip.
